+++ package.json
@@ -18,6 +18,7 @@
     "@types/react-beautiful-dnd": "^13.1.8",
     "@types/react-slick": "^0.23.13",
     "autoprefixer": "^10.4.17",
+    "dotenv": "^16.4.5",
     "framer-motion": "^11.0.3",
     "medium-zoom": "^1.1.0",
     "multer": "^1.4.5-lts.1",
