+++ src/lib/action.ts
@@ -2,11 +2,12 @@
 
 import { Resend } from 'resend'
 import { revalidatePath } from 'next/cache'
-import { Catelog, ESGPdf, Order, Post } from './models'
-import { connectToDb, deleteS3Object, getErrorMessage, getId, getSignedFileUrl, validateString } from './utils'
+import { BatteryPage, Catelog, ESGPdf, Order, Post } from './models'
+import { connectToDb, getErrorMessage, getId, validateString } from './utils'
 import React from 'react'
 import ContactFormEmail from '@/components/customer/contact-us/ContactForm'
 import { S3BucketUrl, batteriesData_admin } from './data'
+import { deleteS3Object, getSignedFileUrl } from './awsUtils'
 
 export const compare = async (username: string, password: string) => {
   if (
@@ -241,5 +242,83 @@ export const sendEmail = async (formData: FormData) => {
 
   return {
     data,
+  }
+}
+
+export const createBatteryPage = async (formData: FormData) => {
+  const { title, itemTitle, itemSubtitle, itemAdvanced, batteryId } = Object.fromEntries(formData)
+  const cateImg: File = formData.get('cateImg') as unknown as File //이미지 데이터
+  const productImgs: File[] = formData.getAll('productImg') as unknown as File[] //항공 ...
+  const productNames: string[] | null = formData.getAll('productName') as unknown as string[] //항공 ...
+  //데이터는 잘 받아옴
+  let imgUrls: string[] = []
+  let signedUrl_cateImg, signedUrl_prodImg
+
+  if (!cateImg) return { success: false, message: '이미지 파일을 찾을 수 없습니다' }
+  console.log('hello')
+  signedUrl_cateImg = await getSignedFileUrl({
+    name: `batteries/${batteriesData_admin[parseInt(batteryId as string)].title}/${title}-category.png`,
+    type: cateImg.type,
+  })
+  const uploadImg = await fetch(signedUrl_cateImg, {
+    method: 'PUT',
+    body: cateImg,
+    headers: { 'Content-type': cateImg.type },
+  })
+  if (uploadImg.status != 200) {
+    return { succes: false, message: '대표 이미지를 저장하는 데에 실패했습니다.' }
+  }
+
+  for (let i = 0; i < productImgs.length; i++) {
+    signedUrl_prodImg = await getSignedFileUrl({
+      name: `batteries/${batteriesData_admin[parseInt(batteryId as string)].title}/${title}/${productNames[i]}`,
+      type: productImgs[i].type,
+    })
+    const uploadImg = await fetch(signedUrl_prodImg, {
+      method: 'PUT',
+      body: productImgs[i],
+      headers: { 'Content-type': productImgs[i].type },
+    })
+    if (uploadImg.status != 200) {
+      return { succes: false, message: '적용제품들의 이미지를 저장하는 데에 실패했습니다.' }
+    } else {
+      imgUrls.push(signedUrl_prodImg.split('?')[0])
+    }
+  }
+  const products: { id: number; name: string; img: string }[] = productImgs.map(function (img: File, id: number) {
+    return { id, name: productNames[id], img: imgUrls[id] }
+  })
+
+  connectToDb() //MongoDB에 연결
+  if (!signedUrl_cateImg || !signedUrl_prodImg) {
+    return { success: false, message: 'SignedURL 생성을 실패했습니다.' }
+  }
+  const data = {
+    title,
+    itemFile: signedUrl_cateImg.split('?')[0],
+    itemTitle,
+    itemSubtitle: itemSubtitle ? itemSubtitle : ' ',
+    itemAdvanced,
+    products,
+  }
+  const prevData = await BatteryPage.find().exec()
+  console.log(prevData)
+  // const res = await BatteryPage.updateOne({ id: parseInt(batteryId as string) }, { data: [...prevData, data] })
+
+  // console.log(batteriesData_admin[parseInt(batteryId as string)], 'BatteryPage successfully updated! response:', res)
+  return { success: true, message: 'createBatteryPage success' }
+  // } catch (error) {
+  //   return { success: false, message: getErrorMessage(error) }
+  // }
+}
+
+export const fetchPageData = async (id: number) => {
+  try {
+    connectToDb()
+    const batteryPageData = await BatteryPage.findOne({ id })
+    return batteryPageData
+  } catch (err) {
+    console.log(err)
+    throw new Error('Failed to fetch data!')
   }
 }
